Skills Panel can now select multiple attributes easily Skills Panel no longer blindly casts objects to subclasses

+++ data/DicePool.java
@@ -29,7 +29,7 @@ public class DicePool {
     dice = new TreeMap<>();
     for (Dice d: Dice.values()) {
       Integer numberOfDice = pool.get(d);
-      dice.put(d, Objects.requireNonNullElse(numberOfDice, (Integer) 0));
+      dice.put(d, Objects.requireNonNullElse(numberOfDice, 0));
     }
   }
